- add filter by software language - new logo - remove old/obsolete design
I recognized that I am not doing the filter by license/language the right
way(tm). Currently I simply grab all sw-objects and extract unique licenses
from them. Although this does not require changes to the database, one
should infact have a license model and add a foreign key to the sw object.
This will scale better when we have more than just a couple of thousand sw
objects...

but hey lets do it when we approach a couple of hundred!