Another db change to get the usage statistics in.
I've decided to in addition to the SoftwareRating and Statistic models to also
have float/int fields in the Software model itself. They still need updating
whenever something is changed in the votes, but will allow for extremely fast
access (no need to first collect all votes, sum them up and divide by nr but
immediately available). This will be future proof and scale to >10,000
submissions :)

I intend to do these rewrites today. One thing, how do we count downloads? Is a
visit to the homepage a download or a site visit? Is a bibtex item download for
the project/ project paper a download or not ?

Anyway I think we will be done wednesday latest with all that...