Extremely intrusive change, that introduces revisions the right way(tm):
Now we have Software and Revision models. Software only holds the name and the
user of the software + statistics. Revisions hold all the rest that used to be
in the Software model plus a key to the software.

This split was necessary as comments and ratings are not revision specific but
static over all software versions.

The software submission form grew a bit (we now have fields such as changes and
thumbnail). While submitting the thumbnail should work already it is not yet
fully integrated (and we should generate thumbnails from screenshots if no
thumbnail is available). Volunteers welcome.

The r-cran script was updated to match the revision/software changes.

The software detail page now shows all possible software revisions. It is also
possible to still look and update the description of older revisions. To this end
we have /revision/view/<id> and /software/view/<id> urls.

I would really like this version to go life ASAP, as it also contains some spam
countermeasures. However we need to write a script first that converts the old
db into the new format. So in case you can think of anything that requires
changes to the db we should even include it now.